py185/01_intro/my_films/pyproject.toml
[tool.poetry]
name = "my-films"
version = "0.1.0"
description = ""
authors = ["SValenziano <senorvalenz@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
psycopg2-binary = "^2.9.10"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
